move some methods around
- move "get_desktop" from utils back to resources
- move "is_stable_version" from utils.updates to utils
- move "check_version" from utils.updates to utils
