+++ maestral/gui/resources/__init__.py
@@ -35,12 +35,36 @@ THEME_DARK = "dark"
 THEME_LIGHT = "light"
 
 
+def get_desktop():
+    """
+    Determines the current desktop environment. This is used for instance to decide
+    which keyring backend is preferred to store the auth token.
+
+    :returns: "gnome", "kde", "xfce", "cocoa", "" or any other string if the desktop
+        $XDG_CURRENT_DESKTOP if the desktop environment is not known to us.
+    :rtype: str
+    """
+
+    if platform.system() == "Linux":
+        current_desktop = os.environ.get("XDG_CURRENT_DESKTOP", "").lower()
+        desktop_session = os.environ.get("GDMSESSION", "").lower()
+
+        for desktop in ("gnome", "kde", "xfce", ""):
+            if desktop in current_desktop or desktop in desktop_session:
+                return desktop
+
+        return current_desktop
+
+    elif platform.system() == "Darwin":
+        return "cocoa"
+
+
 DESKTOP = get_desktop()
 
 
 def get_native_item_icon(item_path):
 
-    if not os.path.exists(item_path):
+    if not osp.exists(item_path):
         # fall back to default file icon
         return get_native_file_icon()
     else:
@@ -61,15 +85,15 @@ def get_native_file_icon():
 def get_system_tray_icon(status, color=None, geometry=None):
     assert status in ("idle", "syncing", "paused", "disconnected", "error")
 
-    gnome_version = __get_gnome_version()
+    gnome_version = _get_gnome_version()
     is_gnome3 = gnome_version is not None and gnome_version[0] >= 3
 
     if DESKTOP == "gnome" and is_gnome3:
         # use symbolic SVG icons created for Gnome 3
         icon_theme_paths = QtGui.QIcon.themeSearchPaths()
-        maestral_icon_path = os.path.join(_root, "icon-theme-gnome")
+        maestral_icon_path = osp.join(_root, "icon-theme-gnome")
         if maestral_icon_path not in icon_theme_paths:
-            icon_theme_paths += [os.path.join(_root, "icon-theme-gnome")]
+            icon_theme_paths += [osp.join(_root, "icon-theme-gnome")]
         QtGui.QIcon.setThemeSearchPaths(icon_theme_paths)
         icon = QtGui.QIcon.fromTheme("menubar_icon_{}-symbolic".format(status))
     elif DESKTOP == "cocoa":
@@ -105,9 +129,9 @@ def statusBarTheme(icon_geometry=None):
     # see if we can trust returned pixel colors
     # (work around for a bug in Qt with KDE where all screenshots return black)
 
-    c0 = __pixel_at(10, 10)
-    c1 = __pixel_at(300, 400)
-    c2 = __pixel_at(800, 800)
+    c0 = _pixel_at(10, 10)
+    c1 = _pixel_at(300, 400)
+    c2 = _pixel_at(800, 800)
 
     if not c0 == c1 == c2 == (0, 0, 0):  # we can trust pixel colors from screenshots
 
@@ -136,7 +160,7 @@ def statusBarTheme(icon_geometry=None):
             py = icon_geometry.bottom()
 
         # get pixel luminance from icon corner or status bar
-        pixel_rgb = __pixel_at(px, py)
+        pixel_rgb = _pixel_at(px, py)
         lum = rgb_to_luminance(*pixel_rgb)
 
         return THEME_LIGHT if lum >= 0.4 else THEME_DARK
@@ -169,14 +193,14 @@ def rgb_to_luminance(r, g, b, base=256):
     return (0.2126*r + 0.7152*g + 0.0722*b)/base
 
 
-def __get_gnome_version():
+def _get_gnome_version():
     gnome3_config_path = "/usr/share/gnome/gnome-version.xml"
     gnome2_config_path = "/usr/share/gnome-about/gnome-version.xml"
 
     xml = None
 
     for path in (gnome2_config_path, gnome3_config_path):
-        if os.path.isfile(path):
+        if osp.isfile(path):
             try:
                 with open(path, "r") as f:
                     xml = f.read()
@@ -194,7 +218,7 @@ def __get_gnome_version():
         return None
 
 
-def __pixel_at(x, y):
+def _pixel_at(x, y):
     """
     Returns (r, g, b) color code for a pixel with given coordinates (each value is in
     0..256 limits)
